feat/quirks: extract quirks from hunspell, add allow_concatenation
Closes #82
